# SE4458-midterm/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SE4458-midterm, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/Ticket/mobileApp/query": {
      "get": {
        "tags": [
          "Ticket"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "query",
            "name": "from",
            "type": "string"
          },
          {
            "in": "query",
            "name": "to",
            "type": "string"
          },
          {
            "in": "query",
            "name": "numberOfPeople",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Ticket/mobileApp/buy": {
      "post": {
        "tags": [
          "Ticket"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TicketPurchaseRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Ticket/mobileApp/checkin": {
      "post": {
        "tags": [
          "Ticket"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CheckInRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Ticket/Admin/insertFlight": {
      "post": {
        "tags": [
          "Ticket"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FlightInsertRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Ticket/Admin/reportFlights": {
      "post": {
        "tags": [
          "Ticket"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "pageNumber",
            "type": "integer",
            "format": "int32",
            "default": 1
          },
          {
            "in": "query",
            "name": "pageSize",
            "type": "integer",
            "format": "int32",
            "default": 10
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FlightReportRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Authentication/Values": {
      "get": {
        "tags": [
          "Values"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "CheckInRequest": {
      "type": "object",
      "properties": {
        "flightNumber": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "passengerName": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "FlightInsertRequest": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "fromLoc": {
          "type": "string"
        },
        "toLoc": {
          "type": "string"
        },
        "availableSeats": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "flightNumber": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "FlightReportRequest": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "fromLoc": {
          "type": "string"
        },
        "toLoc": {
          "type": "string"
        },
        "availableSeats": {
          "format": "int32",
          "type": "integer"
        },
        "minPrice": {
          "format": "double",
          "type": "number"
        },
        "maxPrice": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "TicketPurchaseRequest": {
      "type": "object",
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        },
        "passengerName": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}